Support extraction of $gettext from vue <script> tags
