+++ src/javascript-extract.spec.js
@@ -1,13 +1,13 @@
-const { expect }  = require('chai');
+const { expect } = require('chai');
 
-const fixtures    = require('./test-fixtures.js');
+const fixtures = require('./test-fixtures.js');
 const jsExtractor = require('./javascript-extract.js');
 
 
 describe('Javascript extractor object', () => {
   describe('Extraction of localized strings', () => {
     it('should extract strings from the script', () => {
-      const filename         = fixtures.VUE_COMPONENT_FILENAME;
+      const filename = fixtures.VUE_COMPONENT_FILENAME;
       const extractedStrings = jsExtractor.extractStringsFromJavascript(
         filename,
         fixtures.VUE_COMPONENT_EXPECTED_PROCESSED_SCRIPT_TAG
@@ -15,7 +15,7 @@ describe('Javascript extractor object', () => {
 
       expect(extractedStrings.length).to.be.equal(2);
 
-      const firstString  = extractedStrings[0];
+      const firstString = extractedStrings[0];
       const secondString = extractedStrings[1];
 
       expect(firstString.text).to.be.equal('Hello there!');
